# PERF: Parse certain dates in Cython instead of falling back to dateutil.parse