… doc/source/whatsnew/v0.25.0.rst
… pandas/tests/io/parser/test_parse_dates.py