This tremendously speeds up parsing of certain date patterns: `MM/YYYY`, `MM/DD/YYYY` and `DD/MM/YYYY` (where separator is one of ` `, `/`, `-`, and `\`), as previously they were parsed using `dateutil.parse` which is implemented in Python.

- [x] closes N/A
- [x] tests added / passed
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [x] whatsnew entry