@pep8speaks 2019-03-29
Hello @vnlitvin! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`asv_bench/benchmarks/io/csv.py`](https://github.com/pandas-dev/pandas/blob/72a6ed8cfaffb40491a28823955509cf755d78d6/asv_bench/benchmarks/io/csv.py):

> [Line 240:14](https://github.com/pandas-dev/pandas/blob/72a6ed8cfaffb40491a28823955509cf755d78d6/asv_bench/benchmarks/io/csv.py#L240): [E241](https://duckduckgo.com/?q=pep8%20E241) multiple spaces after ':'