@vnlitvin 2019-03-29
Performance benefits, shown by `asv continuous -f 1.01 upstream/master parse-slashed-date -e -b io.csv.ReadCSVParseSpecialDate -a sample_time=1 -a warmup_time=1`:

before|after|ratio|test name
-|-|-|-
[1d4c89f4]|[72a6ed8c]||
**master**|**parse-slashed-date**||
721±10ms|44.3±2ms|0.06|io.csv.ReadCSVParseSpecialDate.time_read_special_date('mY')
493±4ms|17.9±0.3ms|0.04|io.csv.ReadCSVParseSpecialDate.time_read_special_date('mdY')