@jbrockmendel 2019-03-29
I agree with @mroeschke about creep scope.  Do we have strong reason to believe that performance on this particular set of formats is a major pain point for users?