@vnlitvin 2019-03-30
> I agree with @mroeschke about creep scope. Do we have strong reason to believe that performance on this particular set of formats is a major pain point for users?

As far as I can tell, `dateutil` is pure-Python by design, so it would be really hard to integrate this there.
I don't see why this date format should be treated differently than ISO8601 one (our parsing is definitely way easier to maintain than that).

As for being a pain point - this format is quite widely used in certain real-world data sets, like Mortgage or NYC Taxi Rides, so yes, it _is_ a pain point in some cases (Mortgage dataset that takes dozens of minutes to be loaded _because_ of this format being parsed slowly is the reality I measured).