@jbrockmendel 2019-03-30
> As far as I can tell, dateutil is pure-Python by design, so it would be really hard to integrate this there.

dateutil is pure-python, but it also has lots of room for performance improvement.  Also IIRC @pganssle has been considering adding C/cython/rust backend(s).

> I don't see why this date format should be treated differently than ISO8601 one (our parsing is definitely way easier to maintain than that)

That logic can apply to many formats, most of which we don't want to be responsible for internally.

> Mortgage dataset that takes dozens of minutes to be loaded because of this format being parsed slowly

I think if you pass the `format` kwarg to `to_datetime` it should use `array_strptime`, which I expect would be performant.