@vnlitvin 2019-04-01
> I think if you pass the `format` kwarg to `to_datetime` it should use `array_strptime`, which I expect would be performant.

There's no direct way of passing `format` argument to `to_datetime()` from `read_csv`. I have prototyped passing the argument to run a benchmark, though, and measured the performance.
It is performing roughly the same (in the margin of error), but it requires the caller of `read_csv` to know date format beforehand (including knowing the separator, so if different separators are for some reason used in different rows this would fail to parse correctly).

date format|`format` + `array_strptime`|proposed PR
-|-|-
mY |     43.9±0.4ms   |43.5±0.3ms
mdY|     16.4±0.08ms |  16.5±0.2ms

So I'm stating that using `format` keyword here could save from implementing our speed ups, yes, but at expense of end user knowing the format of input data, while our PR makes parsing almost equally fast without requiring knowledge from the user (and keeping public API unchanged! 😄  ).