@pganssle 2019-04-01
Does this not need additional tests?

I can't comment on the scoping stuff, but one of the hardest parts of changing any auto-magical interface is that it's *very* hard to tell if the things you are doing are going to break something someone is counting on. Even something like this with a fast-path, you should probably be doing a decent amount of testing of the desired behavior *and* similar, related behavior - hard to even say what that would do.

Since you are trying to replace `dateutil.parser.parse`, this may be an ideal situation to use [hypothesis](https://github.com/HypothesisWorks/hypothesis), because you can test the property "this function returns the same thing as `dateutil`" for all kinds of formats, since you're looking for something that is basically a faster version of `dateutil`.