@vnlitvin 2019-04-01
> Does this not need additional tests?

Great question, it most likely does need tests, we'll add some.