@anmyachev 2019-04-02
I have added tests and fixed bug. Thanks @pganssle