@vnlitvin 2019-04-03
> testing that `dateutil` and `pandas` return the same thing for a wide swathe of dates and formats

Note that this change changes the behaviour for dates like `MM/YYYY` slightly, as `dateutil` parses `MM/YYYY` with day set equal to current day while our patch parses it set to `1`. I think our approach is more consistent with how other abbreviated dates are parsed, see 

https://github.com/pandas-dev/pandas/blob/4814a2875d078d6e9e9300bd1f77af15ef4f0f0a/pandas/_libs/tslibs/parsing.pyx#L195-L298

It never replaces `day` in `_DEFAULT_DATETIME` so day stays `1` there.