@pganssle 2019-04-03
@vnlitvin While I think defaulting to "today" was not a good choice in `dateutil`, the reason I have not changed that is that it would be a backwards-incompatible change.

If you are going to make a backwards-incompatible change to the behavior of the library, that should be **very** prominently documented. This change is only documented as a performance improvement.