@vnlitvin 2019-04-03
> This change is only documented as a performance improvement.

That is because I didn't suspect that `dateutil` would use today's day for some unrelated date :)
We will update `whatsnew` with this change.