@Dr-Irv 2019-04-03
@vnlitvin You wrote:

> There's no direct way of passing format argument to to_datetime() from read_csv.

Yes, there is.  You can use something like the following for a MM/YYYY format:
```
df = pd.read_csv(filename, date_parser = lambda d:pd.to_datetime(d, format="%m/%Y")
```

I believe this will use a C implementation of the parser and be very fast.  Can you test that and compare it to what you have put together to see if there is still a significant difference?