@vnlitvin 2019-04-03
> I believe this will use a C implementation of the parser and be very fast. Can you test that and compare it to what you have put together to see if there is still a significant difference?

Sure, thanks for pointing the way.
Here's the results of measuring stuff:

format|master|to_datetime|proposed PR
-|-|-|-
%m/%Y (3 times)|689±5ms|103±0.8ms|43.3±0.5ms
%m/%d/%Y|457±3ms|37.1±0.3ms|16.6±0.2ms

So this `pd.to_datetime` approach is about 2.2x slower than proposed patch, _plus_ user has to know date format beforehand, _plus_ one cannot have mixed dates (like `15/02/2000` and `03/18/2010`). Both current master and our proposed patch allow for _not_ knowing the format and being able to mix MM/DD and DD/MM (I see no use case for that, but still), but ours is about 20x faster.